pyproject.toml
[project]
name = "adek-ai-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-core>=1.35.0",
    "azure-identity>=1.24.0",
    "azure-search-documents>=11.5.3",
    "black>=25.1.0",
    "chromadb>=1.0.20",
    "docx2txt>=0.9",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.28",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "mypy>=1.17.1",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
    "prometheus-client>=0.22.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "structlog>=25.4.0",
    "torch>=2.8.0",
    "transformers>=4.55.4",
    "unstructured[md]>=0.18.14",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]
